loop optimizations
also cut some overshooting lines

+++ source/src/rendergl.cpp
@@ -1125,7 +1125,7 @@ void gl_drawframe(int w, int h, float changelod, float curfps)
         glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
         disablepolygonoffset(GL_POLYGON_OFFSET_LINE, false);
     }
-    extern vector<short> tagclipcubes;
+    extern vector<int> tagclipcubes;
     tagclipcubes.setsize(0);
 
     extern vector<vertex> verts;
